feat: udistrital/planeacion_cliente#962      refactor_planeacion_formato_mid - Se implementa el nuevo formato para control de errores en el API de planeacion_formato_mid
